refactor: extension 분리, price에 number formatter 적용
