+++ OpenMarket/OpenMarket/Controller/OpenMarketViewController+extension.swift
@@ -55,4 +55,26 @@ extension OpenMarketViewController {
             self.gridView.topAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.topAnchor),
         ])
     }
+    
+    func showSpinner(onView : UIView) {
+        let spinnerView = UIView.init(frame: onView.bounds)
+        spinnerView.backgroundColor = UIColor.init(red: 0.5, green: 0.5, blue: 0.5, alpha: 0.5)
+        let activityIndicatorView = UIActivityIndicatorView.init(style: .large)
+        activityIndicatorView.startAnimating()
+        activityIndicatorView.center = spinnerView.center
+        
+        DispatchQueue.main.async {
+            spinnerView.addSubview(activityIndicatorView)
+            onView.addSubview(spinnerView)
+        }
+        
+        loadingView = spinnerView
+    }
+    
+    func removeSpinner() {
+        DispatchQueue.main.async {
+            self.loadingView?.removeFromSuperview()
+            self.loadingView = nil
+        }
+    }
 }
