+++ OpenMarket/OpenMarket/Model/Extension/Int+extension.swift
@@ -1,0 +1,17 @@
+//
+//  Int+extension.swift
+//  OpenMarket
+//
+//  Created by groot, bard on 2022/07/19.
+//
+
+import Foundation
+
+extension Int {
+    func formatNumber() -> String {
+        let numberFormatter = NumberFormatter()
+        numberFormatter.numberStyle = .decimal
+        
+        return numberFormatter.string(for: self) ?? ""
+    }
+}
